Desktop/Personal Projects/Secure Collaboration Platform/secure-collab-starter/secure-collab/backend/app/deps.py
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
import jwt, os
from .auth import JWT_AUD, JWT_ISS

security = HTTPBearer()
JWT_SECRET = os.getenv("JWT_SECRET", "dev_secret_change_me")

def get_current_claims(creds: HTTPAuthorizationCredentials = Depends(security)):
    token = creds.credentials
    try:
        payload = jwt.decode(token, JWT_SECRET, algorithms=["HS256"], audience=JWT_AUD, issuer=JWT_ISS)
        return payload
    except Exception:
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid token")
