Desktop/Personal Projects/Secure Collaboration Platform/secure-collab-starter/secure-collab/backend/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
python-dotenv==1.0.1
